+++ pulse-binding/CHANGELOG.md
@@ -1,3 +1,7 @@
+# [unreleased]
+
+ * Added PA v14 support (API additions)
+
 # 2.16.3 (November 21st, 2020)
 
  * Trivial documentation fixes.