+++ pulse-binding/CHANGELOG.md
@@ -1,9 +1,10 @@
-# [unreleased]
+# 2.17.0 (November 24th, 2020)
 
  * Added PA v14 support (API additions)
  * Mainloop (standard): Changed the `prepare()` method to take an `Option<MicroSeconds>` instead of
    `Option<i32>`.
  * Renamed the error `Code::Io` enum variant to `Code::IO`.
+ * Updated dependencies
 
 # 2.16.3 (November 21st, 2020)